app/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiohttp-swagger==1.0.16
aiosignal==1.3.1
asgiref==3.7.2
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
colorama==0.4.6
commonmark==0.9.1
cryptography==38.0.4
dict2xml==1.7.3
dill==0.3.7
distro==1.8.0
Django==4.2.7
django-environ==0.11.2
f5-icontrol-rest==1.3.13
frozenlist==1.4.0
future==0.18.2
gitdb==4.0.11
GitPython==3.1.40
grpcio==1.59.3
gunicorn==22.0.0
idna==3.4
Jinja2==3.1.2
jsonpickle==3.0.2
junit-xml==1.9
junos-eznc==2.6.5
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.4
mypy-extensions==0.4.3
napalm==4.0.0
ncclient==0.6.13
netaddr==0.8.0
netmiko==4.1.2
netutils==1.3.0
nornir==3.3.0
nornir-jinja2==0.2.0
nornir-napalm==0.3.0
nornir-netbox==0.3.0
nornir-netconf==2.0.0
nornir-netmiko==0.2.0
nornir-scrapli==2022.7.30
nornir-utils==0.2.0
ntc-templates==3.1.0
packaging==23.2
pan-os-python==1.8.1
pan-python==0.17.0
paramiko==2.12.0
pathspec==0.11.2
ply==3.11
prettytable==3.9.0
protobuf==4.25.1
psutil==5.9.6
psycopg2-binary==2.9.9
pyasn1==0.4.8
pycparser==2.21
pyeapi==0.8.4
pyftpdlib==1.5.9
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyserial==3.5
pysmi-lextudio==1.1.13
pysnmp-lextudio==5.0.29
pysnmpcrypto==0.0.4
python-dotenv==1.0.1
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0
requests==2.28.1
rest.connector==23.11
rich==12.6.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scp==0.14.4
scrapli==2022.7.30.post1
scrapli-cfg==2022.7.30
scrapli-community==2022.7.30
scrapli-netconf==2022.7.30
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
tenacity==8.1.0
textfsm==1.1.2
tftpy==0.8.0
tqdm==4.66.1
transitions==0.9.0
ttp==0.9.2
ttp-templates==0.3.2
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.12
xmltodict==0.13.0
yamllint==1.33.0
yamlordereddictloader==0.4.0
yang.connector==23.11
yarl==1.9.3
